Massive restructure for `unittest`.
