+++ quetzal/quetzal/datastructures/README.md
@@ -1,0 +1,3 @@
+# The datastructures
+
+Note about the datastructures.
