+++ quetzal/quetzal/datastructures/__init__.py
@@ -1,0 +1,1 @@
+from .datastructures import *
