+++ quetzal/quetzal/datastructures/datastructures/adt_binary_search_tree.py
