+++ quetzal/quetzal/datastructures/datastructures/adt_circular_linked_list.py
