+++ quetzal/quetzal/datastructures/datastructures/adt_doubly_linked_list.py
