+++ quetzal/quetzal/datastructures/datastructures/adt_stack.py
