+++ quetzal/quetzal/datastructures/run_tests.py
@@ -1,0 +1,16 @@
+#! /bin/python
+
+import unittest
+import tests
+
+def test_adts():
+    suite = unittest.TestSuite()
+    result = unittest.TestResult()
+
+    for test in tests.test_list:
+        suite.addTest(unittest.makeSuite(test))
+
+    runner = unittest.TextTestRunner(verbosity=2)
+    runner.run(suite)
+
+test_adts()
