+++ quetzal/quetzal/datastructures/tests/__init__.py
@@ -1,0 +1,11 @@
+from .test_binary_search_tree import TestBsTree
+from .test_circular_linked_list import TestCircularLinkedList
+from .test_doubly_linked_list import TestDoublyLinkedList
+from .test_hashmap import TestHashmap
+from .test_queue import TestQueue
+from .test_red_black_tree import TestRbTree
+from .test_stack import TestStack
+from .test_two_three_four_tree import TestTwoThreeFourTree
+from .test_two_three_tree import TestTwoThreeTree
+
+test_list = [TestBsTree, TestCircularLinkedList, TestDoublyLinkedList, TestHashmap, TestQueue, TestRbTree, TestStack, TestTwoThreeFourTree, TestTwoThreeTree]
