+++ quetzal/quetzal/datastructures/tests/test_binary_search_tree.py
@@ -1,0 +1,5 @@
+from unittest import TestCase
+
+class TestBsTree(TestCase):
+    def test_1(self):
+        self.assertEqual(1, 1)
