+++ quetzal/quetzal/datastructures/tests/test_circular_linked_list.py
@@ -1,0 +1,5 @@
+from unittest import TestCase
+
+class TestCircularLinkedList(TestCase):
+    def test_1(self):
+        self.assertEqual(1, 1)
