+++ quetzal/quetzal/datastructures/tests/test_doubly_linked_list.py
@@ -1,0 +1,5 @@
+from unittest import TestCase
+
+class TestDoublyLinkedList(TestCase):
+    def test_1(self):
+        self.assertEqual(1, 1)
