+++ quetzal/quetzal/datastructures/tests/test_stack.py
@@ -1,0 +1,5 @@
+from unittest import TestCase
+
+class TestStack(TestCase):
+    def test_1(self):
+        self.assertEqual(1, 1)
