+++ quetzal/quetzal/datastructures/tests/test_two_three_four_tree.py
@@ -1,0 +1,5 @@
+from unittest import TestCase
+
+class TestTwoThreeFourTree(TestCase):
+    def test_1(self):
+        self.assertEqual(1, 1)
