+++ quetzal/quetzal/datastructures/tests/test_two_three_tree.py
@@ -1,0 +1,5 @@
+from unittest import TestCase
+
+class TestTwoThreeTree(TestCase):
+    def test_1(self):
+        self.assertEqual(1, 1)
