+++ quetzal/tests/__init__.py
@@ -1,0 +1,6 @@
+from .test_employee import TestEmployee
+from .test_employeepresent import TestEmployeePresent
+from .test_stock import TestStock
+from .test_user_container import TestUserContainer
+
+test_list = [TestEmployee, TestEmployeePresent, TestStock, TestUserContainer]
